Fix get_gpu_info()
get_cpu_info() called run_cmd() with a shell pipeline which is not supported
by subprocess.check_output() by default. Instead of setting shell=True with
all the quoting fun of running a command through the shell, add an optional
filter argument to filter the lines returned by the command.